+++ dockerfiles/pyforest_sandbox/commands.sh
@@ -1,0 +1,3 @@
+docker build -t 8080labs/pyforest_sandbox . && \
+say "docker update jupypter notebook sandbox ready" && \
+docker run --rm -p 8888:8888 8080labs/pyforest_sandbox