+++ CHANGELOG.rst
@@ -2,6 +2,21 @@
 Changelog
 =========
 
+Version 1.1.0
+=============
+
+- import statements are now automatically added to the first *CODE* cell when a package is used in the notebook
+
+Version 1.0.0
+=============
+
+- import statements are now automatically added to the first cell when a package is used in the notebook
+
+Version 0.1.4
+=============
+
+- Users can now add own lazy imports by writing their import statements to ~/.pyforest/user_imports.py.
+
 Version 0.1
 ===========